refactor database plugin sdk and prepare for ent database plugin sdk development
