+++ sdk/database/dbplugin/v5/grpc_database_plugin.go
@@ -4,13 +4,7 @@
 package dbplugin
 
 import (
-	"context"
-
 	"github.com/hashicorp/go-plugin"
-	"github.com/hashicorp/vault/sdk/database/dbplugin/v5/proto"
-	"github.com/hashicorp/vault/sdk/helper/pluginutil"
-	"github.com/hashicorp/vault/sdk/logical"
-	"google.golang.org/grpc"
 )
 
 // handshakeConfigs are used to just do a basic handshake between
@@ -37,36 +31,3 @@ var (
 	_ plugin.Plugin     = &GRPCDatabasePlugin{}
 	_ plugin.GRPCPlugin = &GRPCDatabasePlugin{}
 )
-
-func (d GRPCDatabasePlugin) GRPCServer(_ *plugin.GRPCBroker, s *grpc.Server) error {
-	var server gRPCServer
-
-	if d.Impl != nil {
-		server = gRPCServer{singleImpl: d.Impl}
-	} else {
-		// multiplexing is supported
-		server = gRPCServer{
-			factoryFunc: d.FactoryFunc,
-			instances:   make(map[string]Database),
-		}
-
-		// Multiplexing is enabled for this plugin, register the server so we
-		// can tell the client in Vault.
-		pluginutil.RegisterPluginMultiplexingServer(s, pluginutil.PluginMultiplexingServerImpl{
-			Supported: true,
-		})
-	}
-
-	proto.RegisterDatabaseServer(s, &server)
-	logical.RegisterPluginVersionServer(s, &server)
-	return nil
-}
-
-func (GRPCDatabasePlugin) GRPCClient(doneCtx context.Context, _ *plugin.GRPCBroker, c *grpc.ClientConn) (interface{}, error) {
-	client := gRPCClient{
-		client:        proto.NewDatabaseClient(c),
-		versionClient: logical.NewPluginVersionClient(c),
-		doneCtx:       doneCtx,
-	}
-	return client, nil
-}
