+++ sdk/database/dbplugin/v5/grpc_server_test.go
@@ -706,26 +706,7 @@ func (badJSONValue) UnmarshalJSON([]byte) error {
 	return fmt.Errorf("this cannot be unmarshalled from JSON")
 }
 
-var _ Database = fakeDatabase{}
-
-type fakeDatabase struct {
-	initResp InitializeResponse
-	initErr  error
-
-	newUserResp NewUserResponse
-	newUserErr  error
-
-	updateUserResp UpdateUserResponse
-	updateUserErr  error
-
-	deleteUserResp DeleteUserResponse
-	deleteUserErr  error
-
-	typeResp string
-	typeErr  error
-
-	closeErr error
-}
+// fakeDatabase methods
 
 func (e fakeDatabase) Initialize(ctx context.Context, req InitializeRequest) (InitializeResponse, error) {
 	return e.initResp, e.initErr
@@ -751,19 +732,7 @@ func (e fakeDatabase) Close() error {
 	return e.closeErr
 }
 
-var _ Database = &recordingDatabase{}
-
-type recordingDatabase struct {
-	initializeCalls int
-	newUserCalls    int
-	updateUserCalls int
-	deleteUserCalls int
-	typeCalls       int
-	closeCalls      int
-
-	// recordingDatabase can act as middleware so we can record the calls to other test Database implementations
-	next Database
-}
+// recordingDatabase methods
 
 func (f *recordingDatabase) Initialize(ctx context.Context, req InitializeRequest) (InitializeResponse, error) {
 	f.initializeCalls++
@@ -813,9 +782,7 @@ func (f *recordingDatabase) Close() error {
 	return f.next.Close()
 }
 
-type fakeDatabaseWithVersion struct {
-	version string
-}
+// fakeDatabaseWithVersion methods
 
 func (e fakeDatabaseWithVersion) PluginVersion() logical.PluginVersion {
 	return logical.PluginVersion{Version: e.version}
